backend/requirements.txt
fastapi
uvicorn
pydantic
openai
langchain
faiss-cpu
sentence-transformers
pandas
numpy
